# Add a test for overlapping buffers (WAS: Don't unroll outer loop for speed™)