@debris 2018-05-19
already merged #30 . Please rebase if still relevant :)