@dvdplm 2018-05-19
Now it's down to adding a test for overlapping buffers. Merge? :)